Use direct db calculation for uninstall details
Use mongodb to calculate the unintall details instead of
in memory. Since it is much more efficient

+++ graylog2-server/src/test/java/org/graylog2/contentpacks/ContentPackInstallationPersistenceServiceTest.java
@@ -170,13 +170,13 @@ public class ContentPackInstallationPersistenceServiceTest {
     @Test
     @UsingDataSet(loadStrategy = LoadStrategyEnum.CLEAN_INSERT)
     public void countInstallationOfEntityById() {
-        final int countedInstallations1 = persistenceService.countInstallationOfEntityById(ModelId.of("5b1a49eb3d274631fe07c86a"));
+        final long countedInstallations1 = persistenceService.countInstallationOfEntityById(ModelId.of("5b1a49eb3d274631fe07c86a"));
         assertThat(countedInstallations1).isEqualTo(2);
 
-        final int countedInstallations2 = persistenceService.countInstallationOfEntityById(ModelId.of("non-exsistant"));
+        final long countedInstallations2 = persistenceService.countInstallationOfEntityById(ModelId.of("non-exsistant"));
         assertThat(countedInstallations2).isEqualTo(0);
 
-        final int countedInstallations3 = persistenceService.countInstallationOfEntityById(ModelId.of("5b1a49eb3d274631fe07befa"));
+        final long countedInstallations3 = persistenceService.countInstallationOfEntityById(ModelId.of("5b1a49eb3d274631fe07befa"));
         assertThat(countedInstallations3).isEqualTo(1);
     }
 }
